// src/interfaces/users.interface.ts
export interface Users {
  user_id: number;
  first_name: string;
  last_name?: string;
  username: string;
  email?: string;
  phone_number: string;
  profile_picture?: string;
  password?: string;
  address_line_first: string;
  address_line_second?: string;
  city?: string;
  state?: string;
  country?: string;
  pincode?: string;
  aadhaar_verification?: boolean;
  email_verified?: boolean;
  phone_verified?: boolean;
  kyc_verified?: boolean;
  role?: string;
  banned_reason?: string;
  bio?: string;
  timezone?: string;
  niche?: string;
  artworks?: string[];
  email_notifications?: boolean;
  tags?: boolean;
  notes?: boolean;
  certification?: Record<string, any>;
  education?: Record<string, any>;
  experience?: string;
  services?: Record<string, any>;
  previous_works?: Record<string, any>;
  projects_created?: number;
  projects_applied?: number;
  projects_completed?: number;
  hire_count?: number;
  review_id?: number;
  total_earnings?: number;
  total_spent?: number;
  payment_method?: Record<string, any>;
  payout_method?: Record<string, any>;
  bank_account_info?: Record<string, any>;
  account_type?: string;
  availability?: string;
  time_spent?: number;
  account_status?: string;
  is_active?: boolean;
  is_banned?: boolean;
  created_at?: Date;
  updated_at?: Date;
  updated_by?: Date;
  last_login_at?: Date;
  reset_token?: string;
  reset_token_expires?: Date;
  login_attempts?: number;
}